Fazendo algumas correções

+++ ex035.py
@@ -1,6 +1,6 @@
-a = int(input('Informe o comprimento da primeira reta: '))
-b = int(input('Informe o comprimento da segunda reta: '))
-c = int(input('Informe o comprimento da terceira reta: '))
+a = float(input('Informe o comprimento da primeira reta: '))
+b = float(input('Informe o comprimento da segunda reta: '))
+c = float(input('Informe o comprimento da terceira reta: '))
 
 if a + b > c and a + c > b and b + c > a:
     print('---------------------------------------\n'
